Enable Nullable reference types.

+++ Spec/Carna.ConsoleRunner.Spec/Configuration/Options/AssemblyOptionSpec_ApplyOption.cs
@@ -1,53 +1,46 @@
-﻿// Copyright (C) 2017 Fievus
+﻿// Copyright (C) 2022 Fievus
 //
 // This software may be modified and distributed under the terms
 // of the MIT license.  See the LICENSE file for details.
-using System;
-using System.IO;
+namespace Carna.ConsoleRunner.Configuration.Options;
 
-namespace Carna.ConsoleRunner.Configuration.Options
+[Context("Applies the assembly option")]
+class AssemblyOptionSpec_ApplyOption : FixtureSteppable, IDisposable
 {
-    [Context("Applies the assembly option")]
-    class AssemblyOptionSpec_ApplyOption : FixtureSteppable, IDisposable
+    AssemblyOption Option { get; } = new();
+    CarnaRunnerCommandLineOptions Options { get; } = new();
+    CarnaRunnerCommandLineOptionContext Context { get; set; } = default!;
+
+    string? AssemblyFilePath { get; set; }
+
+    public void Dispose()
     {
-        AssemblyOption Option { get; } = new AssemblyOption();
-        CarnaRunnerCommandLineOptions Options { get; } = new CarnaRunnerCommandLineOptions();
-        CarnaRunnerCommandLineOptionContext Context { get; set; }
+        if (File.Exists(AssemblyFilePath)) File.Delete(AssemblyFilePath);
+    }
 
-        string AssemblyFilePath { get; set; }
-
-        public void Dispose()
+    [Example("When an assembly file path exists")]
+    void Ex01()
+    {
+        Given("a context that has an assembly file path that exists", () =>
         {
-            if (File.Exists(AssemblyFilePath))
-            {
-                File.Delete(AssemblyFilePath);
-            }
-        }
+            AssemblyFilePath = Path.GetTempFileName();
+            Context = CarnaRunnerCommandLineOptionContext.Of(AssemblyFilePath);
+        });
+        When("the option is applied", () => Option.Apply(Options, Context));
+        Then("the assembly file path should be added", () =>
+            Options.Assemblies.Count == 1 && Options.Assemblies[0] == AssemblyFilePath
+        );
+    }
 
-        [Example("When an assembly file path exists")]
-        void Ex01()
+    [Example("When an assembly file path does not exist")]
+    void Ex02()
+    {
+        Given("a context that has an assembly file path that does not exist", () =>
         {
-            Given("a context that has an assembly file path that exists", () =>
-            {
-                AssemblyFilePath = Path.GetTempFileName();
-                Context = CarnaRunnerCommandLineOptionContext.Of(AssemblyFilePath);
-            });
-            When("the option is applied", () => Option.Apply(Options, Context));
-            Then("the assembly file path should be added", () =>
-                Options.Assemblies.Count == 1 && Options.Assemblies[0] == AssemblyFilePath
-            );
-        }
-
-        [Example("When an assembly file path does not exist")]
-        void Ex02()
-        {
-            Given("a context that has an assembly file path that does not exist", () =>
-            {
-                AssemblyFilePath = "AssemblyFile.dll";
-                Context = CarnaRunnerCommandLineOptionContext.Of(AssemblyFilePath);
-            });
-            When("the option is applied", () => Option.Apply(Options, Context));
-            Then("the InvalidCommandLineOptionException should be thrown", exc => exc.GetType() == typeof(InvalidCommandLineOptionException));
-        }
+            AssemblyFilePath = "AssemblyFile.dll";
+            Context = CarnaRunnerCommandLineOptionContext.Of(AssemblyFilePath);
+        });
+        When("the option is applied", () => Option.Apply(Options, Context));
+        Then("the InvalidCommandLineOptionException should be thrown", exc => exc.GetType() == typeof(InvalidCommandLineOptionException));
     }
 }
